+++ tests/meltano/cli/test_cli.py
@@ -10,9 +10,7 @@ from meltano.cli import cli
 @pytest.fixture(scope="class")
 def project(test_dir, project_init_service):
     """This fixture returns the non-activated project."""
-    project = project_init_service.init(
-        install_default_plugins=False, create_system_database=False, activate=False
-    )
+    project = project_init_service.init(create_system_database=False, activate=False)
 
     yield project